opt-fast/requirements.txt
accelerate==1.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiomysql==0.2.0
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
cryptography==44.0.0
cycler==0.12.1
datasets==3.2.0
dill==0.3.8
easyocr==1.7.2
fastapi==0.115.8
filelock==3.17.0
flatbuffers==25.1.24
fonttools==4.55.8
frozenlist==1.5.0
fsspec==2024.9.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-documentai==3.1.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lazy_loader==0.4
Levenshtein==0.26.1
llvmlite==0.43.0
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.2.1
ninja==1.11.1.3
numba==0.60.0
numpy==2.0.2
onnxruntime==1.19.2
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.1.1
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyMatting==1.1.13
PyMySQL==1.1.1
pyparsing==3.2.1
pyreadline3==3.5.4
pytesseract==0.3.13
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.12.1
referencing==0.36.2
regex==2024.11.6
rembg==2.0.61
requests==2.32.3
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==3.4.1
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
sympy==1.13.1
tesseract==0.1.3
threadpoolctl==3.5.0
tifffile==2024.8.30
tokenizers==0.21.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.48.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
xxhash==3.5.0
yarl==1.18.3
aiokafka==0.8.0
fuzzywuzzy==0.18.0
boto3==1.34.22
flask==3.0.1
gunicorn==21.2.0
konlpy==0.6.0
